api/test_main.py
import unittest
from flask import json
from your_flask_app_file import app

class TestRecommendationAPI(unittest.TestCase):

    def setUp(self):
        # Set up a test client
        self.app = app.test_client()

    def tearDown(self):
        pass

    def test_kg_recommendations(self):
        # Test the kg_recommendations endpoint
        response = self.app.get('/kg_recommendations?userid=1&limit=5')
        data = json.loads(response.data.decode('utf-8'))

        # Add your assertions based on the expected behavior of the endpoint
        self.assertIn('user_id', data[0])
        self.assertIn('similar_user', data[0])
        # Add more assertions as needed

    def test_lstm_predict(self):
        # Test the lstm-predict endpoint
        data = {'text': 'This is a test text'}
        response = self.app.post('/lstm-predict', json=data)
        data = json.loads(response.data.decode('utf-8'))

        # Add your assertions based on the expected behavior of the endpoint
        self.assertIn('text', data)
        self.assertIn('sentiment', data)
        self.assertIn('probability', data)
        # Add more assertions as needed

    def test_content_based_CF_recommendations(self):
        # Test the content_based_CF_recommendations endpoint
        data = {'movie': 'Inception'}
        response = self.app.post('/content_based_CF_recommendations', json=data)
        data = json.loads(response.data.decode('utf-8'))

        # Add your assertions based on the expected behavior of the endpoint
        self.assertIn('recommendations', data)
        # Add more assertions as needed

    def test_tfrs_predict(self):
        # Test the tfrs_predict endpoint
        data = {'user': 1, 'top_n': 5}
        response = self.app.post('/tfrs_predict', json=data)
        data = json.loads(response.data.decode('utf-8'))

        # Add your assertions based on the expected behavior of the endpoint
        self.assertIn('recommendations', data)
        # Add more assertions as needed

if __name__ == '__main__':
    unittest.main()
